Update to kotlin 1.5.30-dev
closes https://github.com/DrewCarlson/WalletKit-Kotlin/issues/2

+++ src/commonMain/kotlin/Blockchains.kt
@@ -18,13 +18,13 @@ internal object Blockchains {
                 BdbCurrency.Denomination(
                         name = "$name Token INT",
                         code = "${codeLowerCase}i",
-                        decimals = 0,
+                        decimals = 0u,
                         symbol = "${codeLowerCase}i"
                 ),
                 BdbCurrency.Denomination(
                         name = "$name Token",
                         code = codeLowerCase,
-                        decimals = decimals.toInt(),
+                        decimals = decimals,
                         symbol = codeLowerCase
                 )
         )
